# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Lumina Gateway API
  version: 0.1.0
  description: >-
    Lumina Gateway backend interface summary (Swagger-style). This document focuses on
    endpoint coverage and brief functional descriptions with minimal schemas.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  schemas:
    GatewayError:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string
      required: [error]
...
paths:
...
  /v1/chat/completions:
    post:
      summary: OpenAI Chat Completions (non-streaming or streaming)
      description: Validates and routes OpenAI chat completion requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                messages:
                  type: array
                  items:
                    type: object
                stream:
                  type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
        '503':
          description: No provider available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
  /v1/responses:
    post:
      summary: OpenAI Responses API (non-streaming or streaming)
      description: Validates and routes OpenAI Responses requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                input:
                  type: array
                stream:
                  type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
        '503':
          description: No provider available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
...
  /openai/v1/chat/completions:
    post:
      summary: OpenAI passthrough (Chat Completions)
      description: Raw passthrough to upstream /chat/completions with failover before first byte.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Upstream response (raw)
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
  /openai/v1/responses:
    post:
      summary: OpenAI passthrough (Responses API)
      description: Raw passthrough to upstream /responses with failover before first byte.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Upstream response (raw)
          content:
            application/json:
              schema:
                type: object
            text/event-stream:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayError'
...